requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.4
django-appconf==1.0.5
django-auth-ldap==4.1.0
django-cors-headers==3.12.0
django-environ==0.8.1
django-filter==21.1
django-imagekit==4.1.0
django-oauth-toolkit==2.0.0
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
hurry.filesize==0.9
idna==3.3
jwcrypto==1.3.1
Markdown==3.3.7
oauthlib==3.2.0
pilkit==2.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
python-ldap==3.4.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
twython==3.9.1
urllib3==1.26.9
wrapt==1.14.1
